src/requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.3
boto3==1.37.23
botocore==1.37.23
certifi==2025.1.31
exceptiongroup==1.2.2
h11==0.14.0
html5lib==1.1
idna==3.10
jmespath==1.0.1
lxml==5.3.1
numpy==2.2.4
outcome==1.3.0.post0
pandas==2.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
s3transfer==0.11.4
selenium==4.30.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
